Refactor: Remove the validation from the controller and create its own class and static method
