+++ test/controllers/PenaltyDetailsController.test.ts
@@ -8,28 +8,21 @@ import '../../src/controllers/PenaltyDetailsController';
 import { RedisService } from '../../src/services/RedisService';
 import { PenaltyReferenceDetails } from '../../src/models/PenaltyReferenceDetails';
 
+
+const app = createApplication(container => {
+    container
+        .bind(RedisService).toConstantValue(createSubstituteOf<RedisService>(service => {
+            service.ping().returns(true)
+        }))
+});
+
 describe('PenaltyDetailsController', () => {
 
     it('should return 200 when trying to access the penalty-reference page', async () => {
-        const app = createApplication(container => {
-
-            container
-                .bind(RedisService).toConstantValue(createSubstituteOf<RedisService>(service => {
-                    service.ping().returns(true)
-                }))
-        });
         await request(app).get('/penalty-reference').expect(200);
     });
 
     it('should return 200 when posting valid penalty detals', async () => {
-
-        const app = createApplication(container => {
-            container
-                .bind(RedisService).toConstantValue(createSubstituteOf<RedisService>(service => {
-                    service.ping().returns(true)
-                }))
-        });
-
         const penaltyDetails: PenaltyReferenceDetails = {
             penaltyReference: 'A12345678',
             companyNumber: 'SC123123'
